@Lazy 사용하여 순화참조 방지되나 근본적인 해결책은 아니므로 수정 필요
The dependencies of some of the beans in the application context form a cycle:

┌─────┐
|  securityConfig defined in file [G:\study_board\out\production\classes\study\board\config\SecurityConfig.class]
↑     ↓
|  customOAuth2UserService defined in file [G:\study_board\out\production\classes\study\board\oauth2\CustomOAuth2UserService.class]
└─────┘

Action:

Relying upon circular references is discouraged and they are prohibited by default. Update your application to remove the dependency cycle between beans. As a last resort, it may be possible to break the cycle automatically by setting spring.main.allow-circular-references to true.